+++ src/main/java/hu/oe/nik/szfmv/automatedcar/TractionForce.java
@@ -1,4 +1,4 @@
-package hu.oe.nik.szfmv.automatedcar.engine;
+package hu.oe.nik.szfmv.automatedcar;
 
 public class TractionForce {
